Item database adapters dev
